rename entity type files
