+++ src/common/typedefs/Application.ts
